requirements.txt
Flask==3.0.3
gunicorn==22.0.0
Jinja2==3.1.4
google-api-python-client
PyPDF2
python-dotenv
google-generativeai
